Revert "SetHealth now receives a Float, corrigindo um bug que o maxTime estava dando anteriormente"
This reverts commit e9e1f5764612c4f5796a5127c3e4c3c16d07ccab.

+++ Assets/Scripts/HealthBar.cs
@@ -10,13 +10,13 @@ public class HealthBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
 
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         slider.maxValue = health;
         slider.value = health;
